requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
-e git+https://github.com/lgarvey/djangosaml2@dev#egg=djangosaml2
    # via -r requirements.txt
-e git+https://github.com/lgarvey/djangosaml2idp.git@multiple-sp-same-entityid-solution#egg=djangosaml2idp
    # via -r requirements.txt
appdirs==1.4.4
    # via black
arrow==1.0.3
    # via
    #   -r requirements.txt
    #   djangosaml2idp
asgiref==3.3.1
    # via
    #   -r requirements.txt
    #   django
atomicwrites==1.1.5
    # via pytest
attrs==18.1.0
    # via pytest
black==20.8b1
    # via -r requirements-dev.in
cachetools==4.2.1
    # via
    #   -r requirements.txt
    #   google-auth
    #   zenpy
certifi==2020.12.5
    # via
    #   -r requirements.txt
    #   elastic-apm
    #   requests
cffi==1.14.4
    # via
    #   -r requirements.txt
    #   cryptography
chardet==3.0.4
    # via
    #   -r requirements.txt
    #   requests
click==7.1.2
    # via black
coverage==4.5.1
    # via pytest-cov
cryptography==3.3.1
    # via
    #   -r requirements.txt
    #   pyopenssl
    #   pysaml2
defusedxml==0.5.0
    # via
    #   -r requirements.txt
    #   djangosaml2
    #   pysaml2
dj-database-url==0.5.0
    # via -r requirements.txt
django-axes==5.12.0
    # via -r requirements.txt
django-environ==0.4.5
    # via -r requirements.txt
django-filter==2.4.0
    # via -r requirements.txt
django-ipware==3.0.2
    # via
    #   -r requirements.txt
    #   django-axes
django-oauth-toolkit==1.3.3
    # via -r requirements.txt
django==3.1.5
    # via
    #   -r requirements.txt
    #   django-axes
    #   django-filter
    #   django-oauth-toolkit
    #   djangorestframework
    #   djangosaml2
    #   djangosaml2idp
djangorestframework==3.12.2
    # via -r requirements.txt
elastic-apm==6.0.0
    # via -r requirements.txt
elementpath==2.1.3
    # via
    #   -r requirements.txt
    #   xmlschema
factory-boy==2.11.1
    # via -r requirements-dev.in
faker==0.8.17
    # via factory-boy
flake8-blind-except==0.1.1
    # via -r requirements-dev.in
flake8-debugger==3.1.0
    # via -r requirements-dev.in
flake8-import-order==0.18
    # via -r requirements-dev.in
flake8-print==3.1.0
    # via -r requirements-dev.in
flake8-quotes==1.0.0
    # via -r requirements-dev.in
flake8==3.5.0
    # via
    #   -r requirements-dev.in
    #   flake8-debugger
    #   flake8-print
    #   flake8-quotes
freezegun==0.3.10
    # via -r requirements-dev.in
google-api-core==1.25.1
    # via
    #   -r requirements.txt
    #   google-api-python-client
google-api-python-client==1.12.8
    # via -r requirements.txt
google-auth-httplib2==0.0.4
    # via
    #   -r requirements.txt
    #   google-api-python-client
google-auth==1.24.0
    # via
    #   -r requirements.txt
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
googleapis-common-protos==1.52.0
    # via
    #   -r requirements.txt
    #   google-api-core
hawk-server==0.0.12
    # via -r requirements.txt
httplib2==0.18.1
    # via
    #   -r requirements.txt
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==2.7
    # via
    #   -r requirements.txt
    #   requests
importlib-resources==5.1.0
    # via
    #   -r requirements.txt
    #   pysaml2
lxml==4.2.4
    # via -r requirements-dev.in
mccabe==0.6.1
    # via flake8
mohawk==1.1.0
    # via -r requirements-dev.in
more-itertools==4.3.0
    # via pytest
mypy-extensions==0.4.3
    # via black
oauth2client==4.1.3
    # via -r requirements.txt
oauthlib==3.1.0
    # via
    #   -r requirements.txt
    #   django-oauth-toolkit
pathspec==0.8.1
    # via black
pluggy==0.7.1
    # via pytest
protobuf==3.14.0
    # via
    #   -r requirements.txt
    #   google-api-core
    #   googleapis-common-protos
psycopg2-binary==2.8.6
    # via -r requirements.txt
py==1.5.4
    # via pytest
pyasn1-modules==0.2.8
    # via
    #   -r requirements.txt
    #   google-auth
    #   oauth2client
pyasn1==0.4.8
    # via
    #   -r requirements.txt
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pycodestyle==2.3.1
    # via
    #   flake8
    #   flake8-debugger
    #   flake8-import-order
    #   flake8-print
pycparser==2.20
    # via
    #   -r requirements.txt
    #   cffi
pyflakes==1.6.0
    # via flake8
pyopenssl==20.0.1
    # via
    #   -r requirements.txt
    #   pysaml2
pysaml2==6.5.1
    # via
    #   -r requirements.txt
    #   djangosaml2
    #   djangosaml2idp
pytest-cov==2.5.1
    # via -r requirements-dev.in
pytest-django==3.3.3
    # via -r requirements-dev.in
pytest-env==0.6.2
    # via -r requirements-dev.in
pytest-mock==1.10.0
    # via -r requirements-dev.in
pytest-sugar==0.9.1
    # via -r requirements-dev.in
pytest==3.7.1
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-mock
    #   pytest-sugar
python-dateutil==2.8.1
    # via
    #   -r requirements.txt
    #   arrow
    #   faker
    #   freezegun
    #   pysaml2
    #   zenpy
pytz==2021.1
    # via
    #   -r requirements.txt
    #   django
    #   djangosaml2idp
    #   google-api-core
    #   pysaml2
    #   zenpy
raven==6.10.0
    # via -r requirements.txt
regex==2021.3.17
    # via black
requests==2.20.1
    # via
    #   -r requirements.txt
    #   django-oauth-toolkit
    #   google-api-core
    #   pysaml2
    #   zenpy
rsa==4.7
    # via
    #   -r requirements.txt
    #   google-auth
    #   oauth2client
six==1.15.0
    # via
    #   -r requirements.txt
    #   cryptography
    #   faker
    #   flake8-print
    #   freezegun
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   mohawk
    #   more-itertools
    #   oauth2client
    #   protobuf
    #   pyopenssl
    #   pysaml2
    #   pytest
    #   python-dateutil
    #   zenpy
sqlparse==0.4.1
    # via
    #   -r requirements.txt
    #   django
termcolor==1.1.0
    # via pytest-sugar
text-unidecode==1.2
    # via faker
toml==0.10.2
    # via black
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via
    #   -r requirements.txt
    #   arrow
    #   black
uritemplate==3.0.1
    # via
    #   -r requirements.txt
    #   google-api-python-client
urllib3==1.24.3
    # via
    #   -r requirements.txt
    #   elastic-apm
    #   requests
waitress==1.4.3
    # via -r requirements.txt
whitenoise==5.2.0
    # via -r requirements.txt
xmlschema==1.4.2
    # via
    #   -r requirements.txt
    #   pysaml2
zenpy==2.0.23
    # via -r requirements.txt
zipp==3.4.0
    # via
    #   -r requirements.txt
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
